chore: reorganizing cpfcnpj folder
